Windows build: Created build.bat, a batch script that will create a virtualenv for the project on the local machine, and use pip to install packages in requirements.txt
Also created some python scripts to ease with git operations. Before pushing, the prepgitpush.py script will move all files in the app/ directory back to a stage/ directory it creates, and delete the app/ directory, as well as any directories virtualenv creates.

The inverse is true for postgitpull.py, which will (after the build script) create an app/ directory and move files from stage/ to it, then delete the stage/ directory.

DEVELOP IN THE APP DIRECTORY TO BE SURE COMMITS ARE PUSHED PROPERLY!